improve the queries: used parameteried queries and move the long ones to a separate file
