+++ app/airflow/dags/libs/find_R_concepts_to_reuse.py
@@ -6,6 +6,11 @@ from libs.utils import (
 )
 from airflow.providers.postgres.hooks.postgres import PostgresHook
 import logging
+from libs.queries import (
+    find_reusing_value_query,
+    find_reusing_field_query,
+    find_object_id_query,
+)
 
 # PostgreSQL connection hook
 pg_hook = PostgresHook(postgres_conn_id="postgres_db_conn")
@@ -23,7 +28,7 @@ def delete_R_concepts(**kwargs) -> None:
         validated_params = pull_validated_params(kwargs, "validate_params")
         table_id = validated_params["table_id"]
 
-        delete_reused_concepts = f"""
+        delete_reused_concepts = """
             DELETE FROM mapping_scanreportconcept
             WHERE creation_type = 'R' 
             AND (
@@ -31,7 +36,7 @@ def delete_R_concepts(**kwargs) -> None:
                 (content_type_id = 22 AND object_id IN (
                     SELECT srf.id 
                     FROM mapping_scanreportfield AS srf
-                    WHERE srf.scan_report_table_id = {table_id}
+                    WHERE srf.scan_report_table_id = %(table_id)s
                 ))
                 OR 
                 -- Delete concepts for values in this table
@@ -39,11 +44,11 @@ def delete_R_concepts(**kwargs) -> None:
                     SELECT srv.id
                     FROM mapping_scanreportvalue AS srv
                     JOIN mapping_scanreportfield AS srf ON srv.scan_report_field_id = srf.id
-                    WHERE srf.scan_report_table_id = {table_id}
+                    WHERE srf.scan_report_table_id = %(table_id)s
                 ))
             );
         """
-        pg_hook.run(delete_reused_concepts)
+        pg_hook.run(delete_reused_concepts % {"table_id": table_id})
 
     except Exception as e:
         logging.error(f"Error in delete_mapping_rules: {str(e)}")
@@ -93,8 +98,8 @@ def find_matching_value(**kwargs):
     )
 
     # Create temp table for reuse concepts
-    create_table_query = f"""
-    CREATE TABLE temp_reuse_concepts_{table_id} (
+    create_table_query = """
+    CREATE TABLE temp_reuse_concepts_%(table_id)s (
         object_id INTEGER,
         matching_value_name TEXT,
         matching_field_name TEXT,
@@ -107,7 +112,7 @@ def find_matching_value(**kwargs):
     );
     """
     try:
-        pg_hook.run(create_table_query)
+        pg_hook.run(create_table_query % {"table_id": table_id})
         logging.info(f"Successfully created temp_standard_concepts_{table_id} table")
     except Exception as e:
         logging.error(
@@ -122,65 +127,15 @@ def find_matching_value(**kwargs):
         "AND eligible_sr_concept.creation_type != 'V'" if field_vocab_pairs else ""
     )
 
-    find_reusing_value_query = f"""
-    INSERT INTO temp_reuse_concepts_{table_id} (
-        matching_value_name, matching_field_name, content_type_id, source_concept_id, source_scanreport_id
-    )
-    SELECT DISTINCT 
-        sr_value.value, 
-        sr_field.name,
-        23,
-        eligible_sr_concept.concept_id,
-        -- TODO: add eligible_sr_concept.standard_concept_id here
-        eligible_scan_report.id
-    FROM mapping_scanreportvalue AS sr_value
-    JOIN mapping_scanreportfield AS sr_field 
-        ON sr_value.scan_report_field_id = sr_field.id
-    JOIN mapping_scanreporttable AS sr_table 
-        ON sr_field.scan_report_table_id = {table_id}
-
-    -- Join to eligible values in other eligible scan reports
-    JOIN mapping_scanreportvalue AS eligible_sr_value 
-        ON sr_value.value = eligible_sr_value.value       -- Matching value's name
-        AND (
-            sr_value.value_description = eligible_sr_value.value_description
-            OR (sr_value.value_description IS NULL AND eligible_sr_value.value_description IS NULL)
-        )                                                 -- Matching value's description.
-    JOIN mapping_scanreportfield AS eligible_sr_field 
-        ON eligible_sr_value.scan_report_field_id = eligible_sr_field.id
-        AND eligible_sr_field.name = sr_field.name        -- Matching value's field name
-    JOIN mapping_scanreporttable AS eligible_sr_table 
-        ON eligible_sr_field.scan_report_table_id = eligible_sr_table.id
-    JOIN mapping_scanreport AS eligible_scan_report 
-        ON eligible_sr_table.scan_report_id = eligible_scan_report.id
-    JOIN mapping_dataset AS dataset 
-        ON eligible_scan_report.parent_dataset_id = dataset.id
-    JOIN mapping_mappingstatus AS map_status 
-        ON eligible_scan_report.mapping_status_id = map_status.id
-    JOIN mapping_scanreportconcept AS eligible_sr_concept
-        ON eligible_sr_concept.object_id = eligible_sr_value.id
-        AND eligible_sr_concept.content_type_id = 23     -- Value level concepts
-        AND eligible_sr_concept.creation_type != 'R'     -- We don't want to reuse R concepts
-        {exclude_v_concepts_condition}
-
-    WHERE dataset.id = {parent_dataset_id}        -- Other conditions
-    AND dataset.hidden = FALSE
-    AND eligible_scan_report.hidden = FALSE
-    AND map_status.value = 'COMPLETE';
-
-    -- After finding eligible matching value, we need to delete (matching_value_name, standard_concept_id (future) , source_concept_id) duplicates, 
-    --only get the occurence with the lowest source_scanreport_id
-    DELETE FROM temp_reuse_concepts_{table_id} AS temp_table
-    USING temp_reuse_concepts_{table_id} AS temp_table_duplicate
-    WHERE
-        temp_table.matching_value_name = temp_table_duplicate.matching_value_name
-        AND temp_table.source_concept_id = temp_table_duplicate.source_concept_id
-        -- TODO: add standard_concept_id matching check (future) here
-        AND temp_table.content_type_id = 23
-        AND temp_table.source_scanreport_id > temp_table_duplicate.source_scanreport_id;
-    """
     try:
-        pg_hook.run(find_reusing_value_query)
+        pg_hook.run(
+            find_reusing_value_query
+            % {
+                "table_id": table_id,
+                "parent_dataset_id": parent_dataset_id,
+                "exclude_v_concepts_condition": exclude_v_concepts_condition,
+            }
+        )
         logging.info(f"Successfully found reusing concepts at the value level")
     except Exception as e:
         logging.error(f"Failed to find reusing concepts at the value level: {str(e)}")
@@ -225,56 +180,14 @@ def find_matching_field(**kwargs):
         details=f"Finding eligible concepts for reuse at the field level",
     )
 
-    find_reusing_field_query = f"""
-    INSERT INTO temp_reuse_concepts_{table_id} (
-        matching_field_name, matching_table_name, content_type_id, source_concept_id, source_scanreport_id
-    )
-    SELECT DISTINCT 
-        sr_field.name, 
-        sr_table.name,
-        22,
-        eligible_sr_concept.concept_id,
-        eligible_scan_report.id
-    FROM mapping_scanreportfield AS sr_field
-    JOIN mapping_scanreporttable AS sr_table 
-        ON sr_field.scan_report_table_id = {table_id}
-
-    -- Join to eligible fields in other eligible scan reports
-    JOIN mapping_scanreportfield AS eligible_sr_field 
-        ON eligible_sr_field.name = sr_field.name        -- Matching field name
-    JOIN mapping_scanreporttable AS eligible_sr_table 
-        ON eligible_sr_field.scan_report_table_id = eligible_sr_table.id
-        AND eligible_sr_table.name = sr_table.name       -- Matching table name
-    JOIN mapping_scanreport AS eligible_scan_report 
-        ON eligible_sr_table.scan_report_id = eligible_scan_report.id
-    JOIN mapping_dataset AS dataset 
-        ON eligible_scan_report.parent_dataset_id = dataset.id
-    JOIN mapping_mappingstatus AS map_status 
-        ON eligible_scan_report.mapping_status_id = map_status.id
-    JOIN mapping_scanreportconcept AS eligible_sr_concept
-        ON eligible_sr_concept.object_id = eligible_sr_field.id
-        AND eligible_sr_concept.content_type_id = 22     -- Field level concepts
-        AND eligible_sr_concept.creation_type != 'R'     -- We don't want to reuse R concepts
-
-    WHERE dataset.id = {parent_dataset_id}        -- Other conditions
-    AND dataset.hidden = FALSE
-    AND eligible_scan_report.hidden = FALSE
-    AND map_status.value = 'COMPLETE';
-
-    -- After finding eligible matching field, we need to delete (matching_field_name, standard_concept_id (future) , source_concept_id) duplicates, 
-    --only get the occurence with the lowest source_scanreport_id
-    DELETE FROM temp_reuse_concepts_{table_id} AS temp_table
-    USING temp_reuse_concepts_{table_id} AS temp_table_duplicate
-    WHERE
-        temp_table.matching_field_name = temp_table_duplicate.matching_field_name
-        AND temp_table.matching_table_name = temp_table_duplicate.matching_table_name
-        AND temp_table.source_concept_id = temp_table_duplicate.source_concept_id
-        -- TODO: add standard_concept_id matching check (future) here
-        AND temp_table.content_type_id = 22
-        AND temp_table.source_scanreport_id > temp_table_duplicate.source_scanreport_id;
-    """
     try:
-        pg_hook.run(find_reusing_field_query)
+        pg_hook.run(
+            find_reusing_field_query
+            % {
+                "table_id": table_id,
+                "parent_dataset_id": parent_dataset_id,
+            }
+        )
         logging.info(f"Successfully found reusing concepts at the field level")
     except Exception as e:
         logging.error(f"Failed to find reusing concepts at the value level: {str(e)}")
@@ -301,37 +214,11 @@ def find_object_id(**kwargs):
     scan_report_id = validated_params["scan_report_id"]
     table_id = validated_params["table_id"]
 
-    find_object_id_query = f"""
-        UPDATE temp_reuse_concepts_{table_id} AS temp_table
-            SET object_id = 
-                CASE 
-                    WHEN temp_table.content_type_id = 22 THEN  -- For ScanReportField
-                        (SELECT sr_field.id 
-                        FROM mapping_scanreportfield AS sr_field 
-                        JOIN mapping_scanreporttable AS sr_table ON sr_field.scan_report_table_id = sr_table.id
-                        WHERE sr_table.scan_report_id = {scan_report_id} AND sr_table.id = {table_id}
-                        AND sr_field.name = temp_table.matching_field_name
-                        AND sr_table.name = temp_table.matching_table_name
-                        LIMIT 1)
-                    WHEN temp_table.content_type_id = 23 THEN  -- For ScanReportValue
-                        (SELECT sr_value.id 
-                        FROM mapping_scanreportvalue AS sr_value
-                        JOIN mapping_scanreportfield AS sr_field ON sr_value.scan_report_field_id = sr_field.id
-                        JOIN mapping_scanreporttable AS sr_table ON sr_field.scan_report_table_id = sr_table.id
-                        WHERE sr_table.scan_report_id = {scan_report_id} AND sr_table.id = {table_id}
-                        AND sr_value.value = temp_table.matching_value_name
-                        AND sr_field.name = temp_table.matching_field_name
-                        LIMIT 1)
-                    ELSE NULL
-                END
-            WHERE (temp_table.content_type_id = 22 AND temp_table.matching_field_name IS NOT NULL)
-            OR (temp_table.content_type_id = 23 AND temp_table.matching_value_name IS NOT NULL);
-
-        DELETE FROM temp_reuse_concepts_{table_id}
-        WHERE object_id IS NULL;
-    """
     try:
-        pg_hook.run(find_object_id_query)
+        pg_hook.run(
+            find_object_id_query
+            % {"table_id": table_id, "scan_report_id": scan_report_id}
+        )
         logging.info(f"Successfully found object ids for reusing concepts")
     except Exception as e:
         logging.error(f"Failed to find object ids for reusing concepts: {str(e)}")
@@ -351,7 +238,7 @@ def create_reusing_concepts(**kwargs):
     scan_report_id = validated_params["scan_report_id"]
 
     # TODO: when source_concept_id is added to the model SCANREPORTCONCEPT, we need to update the query belowto solve the issue #1006
-    create_concept_query = f"""
+    create_concept_query = """
         -- Insert standard concepts for field values (only if they don't already exist)
         INSERT INTO mapping_scanreportconcept (
             created_at,
@@ -371,14 +258,14 @@ def create_reusing_concepts(**kwargs):
             -- TODO: add standard_concept_id here for the newly creted SR concepts
             temp_reuse_concepts.source_concept_id,
             temp_reuse_concepts.content_type_id -- content_type_id for scanreportvalue
-        FROM temp_reuse_concepts_{table_id} AS temp_reuse_concepts;
+        FROM temp_reuse_concepts_%(table_id)s AS temp_reuse_concepts;
 
         -- Drop the temp table holding the temp reusing concepts data after creating the R concepts
-        DROP TABLE IF EXISTS temp_reuse_concepts_{table_id};
+        DROP TABLE IF EXISTS temp_reuse_concepts_%(table_id)s;
         """
 
     try:
-        pg_hook.run(create_concept_query)
+        pg_hook.run(create_concept_query % {"table_id": table_id})
         logging.info("Successfully created R (Reused) concepts")
         update_job_status(
             scan_report=scan_report_id,
